pyproject.toml
[tool.poetry]
name = "out-back-with-bad-db"
version = "0.1.0"
description = ""
authors = ["Semyon Shikota <toylenium@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.4"
drf-spectacular = "^0.27.1"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
